Blank window, clean rendering loop.

+++ unda/src/input/IEventHandler.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stdint.h>
 
 class IEventHandler
 {
@@ -7,14 +8,15 @@ public:
 	IEventHandler() = default;
 	virtual ~IEventHandler() = default;
 
-	virtual void onKeyDown(unsigned int keyCode, bool isRepeat) {}
-	virtual void onKeyUp(unsigned int keyCode, bool isRepeat) {}
+	virtual void onKeyDown(uint32_t keyCode, bool isRepeat) {}
+	virtual void onKeyUp(uint32_t keyCode, bool isRepeat) {}
 	
-	virtual void onMouseDown(unsigned int mouseButton, unsigned short int numClicks) {}
-	virtual void onMouseUp(unsigned int mouseButton, unsigned short int numClicks) {}
+	virtual void onMouseDown(uint32_t mouseButton, uint8_t numClicks) {}
+	virtual void onMouseUp(uint32_t mouseButton, uint8_t numClicks) {}
 	
-	virtual void onMouseMove(int mousePosX, int mousePosY, int deltaX, int deltaY) {}
+	virtual void onMouseMove(int32_t mousePosX, int32_t mousePosY, int32_t deltaX, int32_t deltaY) {}
 
 private:
+
 };
